Support Clang's -Wconversion -Wfloat-conversion -Wsign-conversion with -Werror

+++ test/bench_sort.cpp
@@ -206,7 +206,7 @@ bool measure_algo(T *elements, std::size_t element_count, std::size_t alg, nanos
    nanosecond_type new_clock = timer.elapsed().wall;
 
    //std::cout << "Cmp:" << order_perf_type::num_compare << " Cpy:" << order_perf_type::num_copy;   //for old compilers without ll size argument
-   std::printf("Cmp:%7.03f Cpy:%8.03f", double(order_perf_type::num_compare)/element_count, double(order_perf_type::num_copy)/element_count );
+   std::printf("Cmp:%7.03f Cpy:%8.03f", double(order_perf_type::num_compare)/double(element_count), double(order_perf_type::num_copy)/double(element_count) );
 
    double time = double(new_clock);
